rename some file to meet upstream guideline
